refactor: restructure package name
