+++ contracts/javascore/xcall/src/main/java/foundation/icon/btp/xcall/interfaces/FeeManage.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package foundation.icon.btp.xcall;
+package foundation.icon.btp.xcall.interfaces;
 
 import foundation.icon.score.client.ScoreClient;
 import score.Address;
